pyproject.toml
[project]
name = "django-sphinx-hosting"
version = "1.5.2"
authors = [
    {name = "Caltech IMSS ADS", email = "imss-ads-staff@caltech.edu"},
]
maintainers = [
    {name = "Chris Malek", email = "cmalek@caltech.edu"},
]
requires-python = ">= 3.11"
description = "Reusable Django app for hosting Sphinx documentation privately."
readme = "README.md"
keywords = [
  "documentation",
  "sphinx",
  "django",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "django>=4.2",
    "django-braces>=1.15.0",
    "django-crispy-forms>=2.3",
    "django-extensions>=3.2.3",
    "django-filter>=24.2",
    "django-haystack>=3.3.0",
    "django-rest-framework-helpers>=8.5.0",
    "django-theme-academy>=0.3.11",
    "django-wildewidgets>=0.16.16",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "crispy-bootstrap5>=2024.10",
    "semver>=3.0.2",
    "humanize>=4.11.0",
    "lxml>=4.9.1",
    "cssselect>=1.2.0",
    "rich>=13.9.4"
]

[project.urls]
Documentation = "https://django-sphinx-hosting.readthedocs.io/en/latest/"
Source = "https://github.com/caltechads/django-sphinx-hosting"
Issues = "https://github.com/caltechads/django-sphinx-hosting/issues"

[tool.uv]
python-preference = "only-system"
dev-dependencies = [
    "django-stubs>=5.1.1",
    "boto3-stubs>=1.34.138",
    "botocore-stubs>=1.34.138",
]


[project.optional-dependencies]
docs = [
  "Sphinx<7",
  "sphinx_rtd_theme == 2.0.0",
  "sphinxcontrib-jsonglobaltoc==0.1.1",
  "sphinxcontrib-images >= 0.9.4",
  "sphinx-terraform",
  "diagrams>=0.23.3",
  "setuptools>=75.1.0",
  "sphinx-click>=6.0.0",
]

[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = [
    "setuptools >= 48",
    "wheel >= 0.29.0",
]
build-backend = "setuptools.build_meta"

[tool.mypy]
implicit_optional = "True"
exclude = "(?x)(^build|sandbox\\/setup\\.py$|migrations\\/.*\\.py$)"

[[tool.mypy.overrides]]
module = "sphinx_hosting.views"
disable_error_code = "misc"

[[tool.mypy.overrides]]
module = "haystack.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sphinx_rtd_theme.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework_helpers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "drf_spectacular.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "braces.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework_expandable.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tabulate.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "environ.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "debug_toolbar.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "wildewidgets.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "crequest.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lxml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "crispy_forms.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_extensions.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_filters.*"
ignore_missing_imports = true

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "S603",   # ignore subprocess calls that do not check return code
    "S607",   # ignore subprocess programs that are not absolute paths
    "TRY003", # external messages in exceptions are too verbose
    "T201",   # ignore print statements
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "SIM102", # combine nested ifs
    "UP006",  # stop bugging me about modern type hints
    "FBT001", # stop bugging me about booleans in function signatures
    "FBT002", # stop bugging me about booleans in function signatures
    "N818",   # stop bugging me about not ending my exceptions with "Error"
    "PT009",  # use regular assert instead of unittest asserts
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"